Make revision into range for log command
If only a single revision is specified and it does not specify a range, the
append coerce it to <revision>:0 so we get the log starting at the revision
rather then just the one log entry.

Also if the first "path" specified looks like a revision and no explicit
revision was specified then treat it as a revision.
This is a convenience because I always forget the use -r
This allow us to use:
  sv log -5 HEAD            # get the five most recent log entries.
  lv log  4715:4700  file.c # get the entries for file.c for the given range.